refactor(daemon, lib): split worktrees engine from service adapter
Extracts the worktrees registry logic into a standalone `crate::worktrees`
module, mirroring the engine/adapter split used by the browser bridge and
Snowflake services. The new `WorktreesRegistry` owns the `Mutex<HashMap>`,
TTL reaping, entry cap/eviction, and core ops (register/heartbeat/unregister/
list/first_folder); `WorktreesService` now routes control-socket ops to the
registry, renders the tray menu/status, and drives the VS Code launcher.

Changes:
- Add `src/worktrees.rs` with `WorktreesRegistry`, `RegisterRequest`, and
  `WindowEntry` types; move all liveness logic (reap, evict_oldest,
  sorted_entries) into the engine
- Update `src/daemon/services/worktrees.rs` to delegate ops to the registry
  and simplify its responsibilities to adapter duties only
- Wire the new module into `src/lib.rs`
- Refactor tests: split registry-level unit tests into `src/worktrees.rs`
  and collapse service tests in `src/daemon/services/worktrees.rs` into
  higher-level integration-style tests (e.g., `handle_routes_ops_and_shapes_payloads`)

Impact: Clarifies architectural intent (engine as a reusable registry, adapter
as a thin daemon integration point), improves testability, and aligns the
worktrees service with the established daemon service patterns.
