+++ src/daemon/services/worktrees.rs
@@ -1,128 +1,46 @@
 //! The worktrees daemon service.
 //!
-//! Maintains the live, authoritative set of repos/worktrees open across *every*
-//! VS Code window, fed by a first-party companion extension that reports from
-//! each window over the daemon's control socket. The resident daemon is the
-//! rendezvous point the per-window extension sandbox cannot replace: each window
-//! can see only its own `workspace.workspaceFolders`, so a single process
-//! aggregating those registrations is the only cross-window source of truth.
-//! See ADR-0040.
+//! A thin adapter that hosts the cross-window [`WorktreesRegistry`] under the
+//! daemon's lifecycle and exposes register/heartbeat/unregister/list over the
+//! control socket, plus a tray submenu with a per-window "focus" action.
 //!
-//! Like the Snowflake service this is a cheap, in-memory adapter — no async
-//! setup, no secret persisted. The registry lives behind a [`std::sync::Mutex`]
-//! that is **never held across an `.await`** (the Snowflake rule); every op is
-//! pure CPU under the lock, so liveness reaping happens inline on each read
-//! rather than from a background task.
+//! All registry state and liveness logic (the `Mutex<HashMap>`, TTL reaping, the
+//! entry cap/eviction) lives in [`crate::worktrees`]; this adapter only routes
+//! ops, renders the menu/status, and drives the VS Code launcher. Like the
+//! Snowflake service it is a cheap, in-memory adapter — no async setup, no
+//! secret persisted.
 
-use std::collections::{BTreeSet, HashMap};
+use std::collections::BTreeSet;
 use std::path::{Path, PathBuf};
 use std::process::{Command, Stdio};
-use std::sync::{Mutex, MutexGuard, PoisonError};
-use std::time::Duration;
 
 use anyhow::{anyhow, bail, Context, Result};
 use async_trait::async_trait;
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 
 use crate::daemon::service::{DaemonService, MenuAction, MenuItem, MenuSnapshot, ServiceStatus};
+use crate::worktrees::{RegisterRequest, WindowEntry, WorktreesRegistry};
 
 /// The worktrees service name (the control-socket routing key).
 pub const SERVICE_NAME: &str = "worktrees";
-
-/// How long a window may go silent before it ages out of the registry. Three
-/// missed ~10s heartbeats; a window that crashed without firing `unregister`
-/// disappears on the next read. The resident process is what makes this
-/// liveness correct — a flat shared file could not reap stale entries.
-const DEFAULT_TTL: Duration = Duration::from_secs(30);
-
-/// Ceiling on live registry entries, so a misbehaving companion flooding
-/// `register` with distinct keys cannot grow daemon memory faster than the TTL
-/// reaps it (#1140). Far above any real window count; when a new key would
-/// exceed it, the longest-silent entry is evicted instead of rejecting the
-/// request — an evicted live window self-heals via the `heartbeat` →
-/// `{known: false}` → re-register path, so `register` stays infallible for the
-/// companion.
-const MAX_WINDOWS: usize = 256;
 
 /// Environment override for the VS Code launcher used by the "focus" tray
 /// action, for when the daemon runs under launchd with a minimal `PATH`.
 const VSCODE_BIN_ENV: &str = "OMNI_DEV_VSCODE_BIN";
 
-/// A `register` request from a companion extension. The companion owns its
-/// `key` (a per-`activate()` UUID) so the daemon never has to reason about
-/// whether `vscode.env.sessionId` is unique per window; everything else is
-/// best-effort metadata.
-#[derive(Debug, Clone, Deserialize)]
-struct RegisterRequest {
-    /// Stable per-window identity, generated by the companion on activation.
-    key: String,
-    /// Absolute paths of the window's workspace folders.
-    #[serde(default)]
-    folders: Vec<PathBuf>,
-    /// Repository root or name, when the window has one.
-    #[serde(default)]
-    repo: Option<String>,
-    /// The window title, for display.
-    #[serde(default)]
-    title: Option<String>,
-    /// The reporting extension-host process id.
-    #[serde(default)]
-    pid: Option<u32>,
-}
-
-/// One open window's live registration. Serialized verbatim into `list` /
-/// `status` payloads; consumers compute age from `last_seen` (RFC 3339).
-#[derive(Debug, Clone, Serialize)]
-struct WindowEntry {
-    /// The companion-owned per-window key.
-    key: String,
-    /// Absolute workspace-folder paths.
-    folders: Vec<PathBuf>,
-    /// Repository root or name, if reported.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    repo: Option<String>,
-    /// Window title, if reported.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    title: Option<String>,
-    /// Reporting extension-host pid, if reported.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pid: Option<u32>,
-    /// When the daemon last heard from this window (register or heartbeat).
-    last_seen: DateTime<Utc>,
-}
-
-/// Hosts the cross-window worktree registry as a [`DaemonService`].
+/// Hosts the cross-window [`WorktreesRegistry`] as a [`DaemonService`].
 pub struct WorktreesService {
-    /// Open windows keyed by their companion-owned `key`.
-    windows: Mutex<HashMap<String, WindowEntry>>,
-    /// How long an entry survives without a heartbeat.
-    ttl: Duration,
+    /// The cross-window registry this adapter routes ops to.
+    registry: WorktreesRegistry,
 }
 
 impl WorktreesService {
-    /// Creates the service with the default liveness TTL. Cheap — no I/O.
+    /// Creates the service with an empty registry. Cheap — no I/O.
     #[must_use]
     pub fn new() -> Self {
         Self {
-            windows: Mutex::new(HashMap::new()),
-            ttl: DEFAULT_TTL,
+            registry: WorktreesRegistry::new(),
         }
-    }
-
-    /// Locks the registry, recovering from a poisoned mutex (a panic in a prior
-    /// critical section must not wedge the whole service).
-    fn lock(&self) -> MutexGuard<'_, HashMap<String, WindowEntry>> {
-        self.windows.lock().unwrap_or_else(PoisonError::into_inner)
-    }
-
-    /// Reaps stale entries, then returns the live set sorted for deterministic
-    /// output. Holds the lock only for pure-CPU work.
-    fn live_entries(&self, now: DateTime<Utc>) -> Vec<WindowEntry> {
-        let mut windows = self.lock();
-        reap(&mut windows, self.ttl, now);
-        sorted_entries(&windows)
     }
 }
 
@@ -139,7 +57,6 @@ impl DaemonService for WorktreesService {
     }
 
     async fn handle(&self, op: &str, payload: Value) -> Result<Value> {
-        let now = Utc::now();
         match op {
             "register" => {
                 let req: RegisterRequest =
@@ -147,56 +64,24 @@ impl DaemonService for WorktreesService {
                 if req.key.trim().is_empty() {
                     bail!("`register` requires a non-empty `key`");
                 }
-                let mut windows = self.lock();
-                reap(&mut windows, self.ttl, now);
-                // Upserts never evict; only a genuinely new key can grow the
-                // map, and never past MAX_WINDOWS.
-                if !windows.contains_key(&req.key) && windows.len() >= MAX_WINDOWS {
-                    evict_oldest(&mut windows);
-                }
-                windows.insert(
-                    req.key.clone(),
-                    WindowEntry {
-                        key: req.key,
-                        folders: req.folders,
-                        repo: req.repo,
-                        title: req.title,
-                        pid: req.pid,
-                        last_seen: now,
-                    },
-                );
+                self.registry.register(req);
                 Ok(json!({ "ok": true }))
             }
             "heartbeat" => {
                 let key = require_key(&payload, "heartbeat")?;
-                let mut windows = self.lock();
-                reap(&mut windows, self.ttl, now);
-                // `known: false` tells a window that started before the daemon —
-                // or survived a daemon restart — to re-`register`, since the
-                // registry is in-memory and has no record of it.
-                let known = match windows.get_mut(key) {
-                    Some(entry) => {
-                        entry.last_seen = now;
-                        true
-                    }
-                    None => false,
-                };
-                Ok(json!({ "known": known }))
+                Ok(json!({ "known": self.registry.heartbeat(key) }))
             }
             "unregister" => {
                 let key = require_key(&payload, "unregister")?;
-                let mut windows = self.lock();
-                let removed = windows.remove(key).is_some();
-                reap(&mut windows, self.ttl, now);
-                Ok(json!({ "removed": removed }))
+                Ok(json!({ "removed": self.registry.unregister(key) }))
             }
-            "list" => Ok(json!({ "windows": self.live_entries(now) })),
+            "list" => Ok(json!({ "windows": self.registry.list() })),
             other => bail!("unknown worktrees op: {other}"),
         }
     }
 
     fn menu(&self) -> MenuSnapshot {
-        let entries = self.live_entries(Utc::now());
+        let entries = self.registry.list();
         let items = if entries.is_empty() {
             vec![MenuItem::Label("No open windows".to_string())]
         } else {
@@ -210,13 +95,11 @@ impl DaemonService for WorktreesService {
 
     async fn menu_action(&self, action_id: &str) -> Result<()> {
         if let Some(key) = action_id.strip_prefix("focus:") {
-            // Resolve the folder under the lock, then drop it before spawning —
-            // never hold the mutex across the process launch.
-            let folder = {
-                let windows = self.lock();
-                windows.get(key).and_then(|e| e.folders.first().cloned())
-            };
-            let folder = folder
+            // The registry resolves the folder under its own lock and clones it
+            // out, so the mutex is never held across the process launch.
+            let folder = self
+                .registry
+                .first_folder(key)
                 .ok_or_else(|| anyhow!("no open window with key {key} (it may have closed)"))?;
             focus_window(&folder)?;
             return Ok(());
@@ -225,7 +108,7 @@ impl DaemonService for WorktreesService {
     }
 
     async fn status(&self) -> ServiceStatus {
-        let entries = self.live_entries(Utc::now());
+        let entries = self.registry.list();
         let repos: BTreeSet<&str> = entries.iter().filter_map(|e| e.repo.as_deref()).collect();
         ServiceStatus {
             name: SERVICE_NAME.to_string(),
@@ -247,39 +130,6 @@ fn require_key<'a>(payload: &'a Value, op: &str) -> Result<&'a str> {
         .get("key")
         .and_then(Value::as_str)
         .ok_or_else(|| anyhow!("`{op}` requires `key`"))
-}
-
-/// Removes entries last seen longer than `ttl` ago. Pure CPU; the caller holds
-/// the registry lock but never `.await`s while holding it.
-fn reap(windows: &mut HashMap<String, WindowEntry>, ttl: Duration, now: DateTime<Utc>) {
-    let max_age = ttl.as_secs() as i64;
-    windows.retain(|_, e| (now - e.last_seen).num_seconds() <= max_age);
-}
-
-/// Removes the entry with the oldest `last_seen` (ties broken by lowest key
-/// for determinism). Called when a `register` of a new key would grow the
-/// registry past [`MAX_WINDOWS`]. Pure CPU under the registry lock, like
-/// [`reap`].
-fn evict_oldest(windows: &mut HashMap<String, WindowEntry>) {
-    let oldest = windows
-        .values()
-        .min_by(|a, b| {
-            a.last_seen
-                .cmp(&b.last_seen)
-                .then_with(|| a.key.cmp(&b.key))
-        })
-        .map(|e| e.key.clone());
-    if let Some(key) = oldest {
-        windows.remove(&key);
-    }
-}
-
-/// Snapshots the registry into a stably-ordered vector (by repo, then key) so
-/// `list`/`status`/`menu` output is deterministic despite `HashMap` ordering.
-fn sorted_entries(windows: &HashMap<String, WindowEntry>) -> Vec<WindowEntry> {
-    let mut entries: Vec<WindowEntry> = windows.values().cloned().collect();
-    entries.sort_by(|a, b| a.repo.cmp(&b.repo).then_with(|| a.key.cmp(&b.key)));
-    entries
 }
 
 /// A short human name for a window: its repo, else its first folder's basename,
@@ -409,6 +259,7 @@ fn resolve_code_binary_from(
 #[allow(clippy::unwrap_used, clippy::expect_used)]
 mod tests {
     use super::*;
+    use chrono::Utc;
 
     fn register_payload(key: &str, repo: Option<&str>, folder: &str) -> Value {
         json!({
@@ -436,250 +287,60 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn list_is_empty_initially() {
+    async fn handle_routes_ops_and_shapes_payloads() {
         let svc = WorktreesService::new();
+        // Empty to start.
         let payload = svc.handle("list", Value::Null).await.unwrap();
         assert_eq!(payload, json!({ "windows": [] }));
-    }
 
-    #[tokio::test]
-    async fn register_then_list_round_trips() {
-        let svc = WorktreesService::new();
+        // register → { ok: true }, then it shows up in list.
         let reply = svc
             .handle("register", register_payload("w1", Some("repo-a"), "/tmp/a"))
             .await
             .unwrap();
         assert_eq!(reply, json!({ "ok": true }));
-
-        let payload = svc.handle("list", Value::Null).await.unwrap();
-        let windows = windows_of(&payload);
+        let windows = windows_of(&svc.handle("list", Value::Null).await.unwrap()).clone();
         assert_eq!(windows.len(), 1);
         assert_eq!(windows[0].get("key").and_then(Value::as_str), Some("w1"));
-        assert_eq!(
-            windows[0].get("repo").and_then(Value::as_str),
-            Some("repo-a")
-        );
         assert!(windows[0].get("last_seen").is_some());
-    }
 
-    #[tokio::test]
-    async fn register_is_idempotent_upsert() {
-        let svc = WorktreesService::new();
-        svc.handle("register", register_payload("w1", Some("repo-a"), "/tmp/a"))
-            .await
-            .unwrap();
-        // Re-registering the same key updates rather than duplicates.
-        svc.handle("register", register_payload("w1", Some("repo-b"), "/tmp/b"))
-            .await
-            .unwrap();
-        let payload = svc.handle("list", Value::Null).await.unwrap();
-        let windows = windows_of(&payload);
-        assert_eq!(windows.len(), 1);
-        assert_eq!(
-            windows[0].get("repo").and_then(Value::as_str),
-            Some("repo-b")
-        );
-    }
-
-    #[tokio::test]
-    async fn register_requires_key() {
-        let svc = WorktreesService::new();
-        assert!(svc.handle("register", json!({})).await.is_err());
-        assert!(svc
-            .handle("register", json!({ "key": "  " }))
-            .await
-            .is_err());
-    }
-
-    #[tokio::test]
-    async fn heartbeat_reports_known_and_unknown() {
-        let svc = WorktreesService::new();
-        // Unknown before registration: the window must re-register.
-        let unknown = svc
-            .handle("heartbeat", json!({ "key": "w1" }))
-            .await
-            .unwrap();
-        assert_eq!(unknown, json!({ "known": false }));
-
-        svc.handle("register", register_payload("w1", None, "/tmp/a"))
-            .await
-            .unwrap();
+        // heartbeat known/unknown.
         let known = svc
             .handle("heartbeat", json!({ "key": "w1" }))
             .await
             .unwrap();
         assert_eq!(known, json!({ "known": true }));
-        assert!(svc.handle("heartbeat", json!({})).await.is_err());
-    }
-
-    #[tokio::test]
-    async fn unregister_removes() {
-        let svc = WorktreesService::new();
-        svc.handle("register", register_payload("w1", None, "/tmp/a"))
+        let unknown = svc
+            .handle("heartbeat", json!({ "key": "nope" }))
             .await
             .unwrap();
+        assert_eq!(unknown, json!({ "known": false }));
+
+        // unregister removes, then repeats as a no-op success.
         let gone = svc
             .handle("unregister", json!({ "key": "w1" }))
             .await
             .unwrap();
         assert_eq!(gone, json!({ "removed": true }));
-        // Removing again is a no-op success.
         let again = svc
             .handle("unregister", json!({ "key": "w1" }))
             .await
             .unwrap();
         assert_eq!(again, json!({ "removed": false }));
+    }
+
+    #[tokio::test]
+    async fn handle_rejects_missing_or_empty_key() {
+        let svc = WorktreesService::new();
+        // register validates a present, non-blank key.
+        assert!(svc.handle("register", json!({})).await.is_err());
+        assert!(svc
+            .handle("register", json!({ "key": "  " }))
+            .await
+            .is_err());
+        // heartbeat/unregister require the key via `require_key`.
+        assert!(svc.handle("heartbeat", json!({})).await.is_err());
         assert!(svc.handle("unregister", json!({})).await.is_err());
-    }
-
-    #[test]
-    fn reap_evicts_only_stale_entries() {
-        let now = Utc::now();
-        let mut windows = HashMap::new();
-        windows.insert(
-            "fresh".to_string(),
-            WindowEntry {
-                key: "fresh".to_string(),
-                folders: vec![],
-                repo: None,
-                title: None,
-                pid: None,
-                last_seen: now - chrono::Duration::seconds(5),
-            },
-        );
-        windows.insert(
-            "stale".to_string(),
-            WindowEntry {
-                key: "stale".to_string(),
-                folders: vec![],
-                repo: None,
-                title: None,
-                pid: None,
-                last_seen: now - chrono::Duration::seconds(120),
-            },
-        );
-        reap(&mut windows, DEFAULT_TTL, now);
-        assert!(windows.contains_key("fresh"));
-        assert!(!windows.contains_key("stale"));
-    }
-
-    /// A minimal entry for cap/eviction tests; only `key` and `last_seen`
-    /// participate in eviction order.
-    fn entry_at(key: &str, last_seen: DateTime<Utc>) -> WindowEntry {
-        WindowEntry {
-            key: key.to_string(),
-            folders: vec![],
-            repo: None,
-            title: None,
-            pid: None,
-            last_seen,
-        }
-    }
-
-    #[test]
-    fn evict_oldest_removes_oldest_with_key_tiebreak() {
-        let now = Utc::now();
-        let mut windows = HashMap::new();
-        windows.insert("young".to_string(), entry_at("young", now));
-        windows.insert(
-            "old-b".to_string(),
-            entry_at("old-b", now - chrono::Duration::seconds(10)),
-        );
-        windows.insert(
-            "old-a".to_string(),
-            entry_at("old-a", now - chrono::Duration::seconds(10)),
-        );
-        // Oldest `last_seen` is shared by two entries; the lowest key loses.
-        evict_oldest(&mut windows);
-        assert!(!windows.contains_key("old-a"));
-        assert!(windows.contains_key("old-b"));
-        assert!(windows.contains_key("young"));
-        // Empty map is a no-op rather than a panic.
-        let mut empty: HashMap<String, WindowEntry> = HashMap::new();
-        evict_oldest(&mut empty);
-        assert!(empty.is_empty());
-    }
-
-    #[tokio::test]
-    async fn register_at_cap_evicts_only_the_oldest() {
-        let svc = WorktreesService::new();
-        // Seed a full registry directly (registering 256 times over the socket
-        // op would work too, but sub-second timestamps may tie; explicit
-        // timestamps make the highest-numbered key unambiguously the oldest).
-        {
-            let mut windows = svc.lock();
-            let base = Utc::now();
-            for i in 0..MAX_WINDOWS {
-                let key = format!("w{i:03}");
-                windows.insert(
-                    key.clone(),
-                    entry_at(&key, base - chrono::Duration::milliseconds(i as i64)),
-                );
-            }
-        }
-        // A new key at the cap displaces exactly the longest-silent entry.
-        svc.handle("register", register_payload("fresh", None, "/tmp/f"))
-            .await
-            .unwrap();
-        let windows = svc.lock();
-        assert_eq!(windows.len(), MAX_WINDOWS);
-        assert!(windows.contains_key("fresh"));
-        assert!(!windows.contains_key(&format!("w{:03}", MAX_WINDOWS - 1)));
-        assert!(windows.contains_key("w000"));
-    }
-
-    #[tokio::test]
-    async fn register_upsert_at_cap_does_not_evict() {
-        let svc = WorktreesService::new();
-        {
-            let mut windows = svc.lock();
-            let base = Utc::now();
-            for i in 0..MAX_WINDOWS {
-                let key = format!("w{i:03}");
-                windows.insert(
-                    key.clone(),
-                    entry_at(&key, base - chrono::Duration::milliseconds(i as i64)),
-                );
-            }
-        }
-        // Re-registering an existing key is an upsert: nothing is displaced,
-        // not even the oldest entry.
-        let oldest = format!("w{:03}", MAX_WINDOWS - 1);
-        svc.handle("register", register_payload(&oldest, Some("r"), "/tmp/a"))
-            .await
-            .unwrap();
-        let windows = svc.lock();
-        assert_eq!(windows.len(), MAX_WINDOWS);
-        assert!(windows.contains_key(&oldest));
-        assert!(windows.contains_key("w000"));
-    }
-
-    #[test]
-    fn sorted_entries_orders_by_repo_then_key() {
-        let now = Utc::now();
-        let mut windows = HashMap::new();
-        for (key, repo) in [("z", "repo-a"), ("a", "repo-b"), ("m", "repo-a")] {
-            windows.insert(
-                key.to_string(),
-                WindowEntry {
-                    key: key.to_string(),
-                    folders: vec![],
-                    repo: Some(repo.to_string()),
-                    title: None,
-                    pid: None,
-                    last_seen: now,
-                },
-            );
-        }
-        let entries = sorted_entries(&windows);
-        let ordered: Vec<(&str, &str)> = entries
-            .iter()
-            .map(|e| (e.key.as_str(), e.repo.as_deref().unwrap()))
-            .collect();
-        assert_eq!(
-            ordered,
-            vec![("m", "repo-a"), ("z", "repo-a"), ("a", "repo-b")]
-        );
     }
 
     #[test]
@@ -715,12 +376,6 @@ mod tests {
             ..base
         };
         assert_eq!(display_name(&rootish), "/");
-    }
-
-    #[test]
-    fn default_constructs_an_empty_service() {
-        let svc = WorktreesService::default();
-        assert!(svc.lock().is_empty());
     }
 
     #[test]
@@ -807,6 +462,13 @@ mod tests {
             .collect();
         assert!(action_ids.contains(&"focus:w1"));
         assert!(action_ids.contains(&"focus:w2"));
+    }
+
+    #[tokio::test]
+    async fn default_constructs_an_empty_service() {
+        let svc = WorktreesService::default();
+        let payload = svc.handle("list", Value::Null).await.unwrap();
+        assert_eq!(payload, json!({ "windows": [] }));
     }
 
     #[tokio::test]
